Re #27007 Convert SansCalculateTransmission into func
Converts SANSCalculateTransmission into a function from a class, as we
do not need to persist any state internally
